Compile Python extensions using CMake
Maintaining consistency between CMake and Python build tools
configuration is difficult and for multi-extension projects builds are
slow. To resolve this for other Cython-based extensions CMake was used
to compile Python binary extensions, specifically the PythonLibs
python_add_module() CMake function.

This leaves us without installation logic; guessing the configured
installation location across platforms and configurations can be
unreliabable; to this end we use Python setuptools (setup.py) to perform
the installation but override the "build_ext" command with an
instruction to copy the library we already generated in the build phase.

We can then install and use Python extensions within a virtual
environment.